chore: clean macro
