+++ src/catalog/mod.rs
@@ -3,5 +3,5 @@
 pub(crate) use self::column::*;
 pub(crate) use self::table::*;
 
-mod column;
-mod table;
+pub mod column;
+pub mod table;
