+++ src/function/numbers.rs
@@ -5,24 +5,85 @@ use crate::errors::DatabaseError;
 use crate::expression::function::table::TableFunctionImpl;
 use crate::expression::function::FunctionSummary;
 use crate::expression::ScalarExpression;
-use crate::table_function;
 use crate::types::tuple::SchemaRef;
 use crate::types::tuple::Tuple;
-use crate::types::value::{DataValue, ValueRef};
+use crate::types::value::DataValue;
 use crate::types::LogicalType;
 use lazy_static::lazy_static;
 use serde::Deserialize;
 use serde::Serialize;
 use std::sync::Arc;
 
-table_function!(Numbers::numbers(LogicalType::Integer) -> [number: LogicalType::Integer] => (|v1: ValueRef| {
-    let num = v1.i32().ok_or_else(|| DatabaseError::NotNull)?;
+lazy_static! {
+    static ref NUMBERS: TableCatalog = {
+        TableCatalog::new(
+            Arc::new("numbers".to_lowercase()),
+            vec![ColumnCatalog::new(
+                "number".to_lowercase(),
+                true,
+                ColumnDesc::new(LogicalType::Integer, false, false, None),
+            )],
+        )
+        .unwrap()
+    };
+}
 
-    Ok(Box::new((0..num)
-        .map(|i| Ok(Tuple {
+#[derive(Debug, Serialize, Deserialize)]
+pub(crate) struct Numbers {
+    summary: FunctionSummary,
+}
+
+impl Numbers {
+    #[allow(unused_mut)]
+    pub(crate) fn new() -> Arc<Self> {
+        let function_name = "numbers".to_lowercase();
+
+        Arc::new(Self {
+            summary: FunctionSummary {
+                name: function_name,
+                arg_types: vec![LogicalType::Integer],
+            },
+        })
+    }
+}
+
+#[typetag::serde]
+impl TableFunctionImpl for Numbers {
+    #[allow(unused_variables, clippy::redundant_closure_call)]
+    fn eval(
+        &self,
+        args: &[ScalarExpression],
+    ) -> Result<Box<dyn Iterator<Item = Result<Tuple, DatabaseError>>>, DatabaseError> {
+        let tuple = Tuple {
             id: None,
-            values: vec![
-                Arc::new(DataValue::Int32(Some(i))),
-            ]
-        }))) as Box<dyn Iterator<Item = Result<Tuple, DatabaseError>>>)
-}));
+            values: Vec::new(),
+        };
+
+        let mut value = args[0].eval(&tuple, &[])?;
+
+        if value.logical_type() != LogicalType::Integer {
+            value = Arc::new(DataValue::clone(&value).cast(&LogicalType::Integer)?);
+        }
+        let num = value.i32().ok_or_else(|| DatabaseError::NotNull)?;
+
+        Ok(Box::new((0..num).map(|i| {
+            Ok(Tuple {
+                id: None,
+                values: vec![Arc::new(DataValue::Int32(Some(i)))],
+            })
+        }))
+            as Box<dyn Iterator<Item = Result<Tuple, DatabaseError>>>)
+    }
+
+    fn output_schema(&self) -> &SchemaRef {
+        NUMBERS.schema_ref()
+    }
+
+    fn summary(&self) -> &FunctionSummary {
+        &self.summary
+    }
+
+    fn table(&self) -> &'static TableCatalog {
+        &NUMBERS
+    }
+}
