+++ tests/macros-test/Cargo.toml
@@ -1,0 +1,11 @@
+[package]
+name    = "macros-test"
+version = "0.4.0"
+edition = "2021"
+
+[dev-dependencies]
+"fnck_sql"  = { path = "../.." }
+lazy_static = { version = "1" }
+serde       = { version = "1", features = ["derive", "rc"] }
+sqlparser   = { version = "0.34", features = ["serde"] }
+typetag     = { version = "0.2" }
